requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
Flask==3.0.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.314
langsmith==0.0.43
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.29.0
sniffio==1.3.0
SQLAlchemy==2.0.22
superpowered-sdk==0.1.22
tenacity==8.2.3
tqdm==4.66.1
twilio==8.9.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.17
Werkzeug==3.0.0
yarl==1.9.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.4.1
asttokens==2.0.8
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
black==22.10.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
cycler==0.11.0
dataclasses-json==0.6.1
debugpy==1.6.3
decorator==5.1.1
Django==3.2.7
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.1.1
Flask==2.0.1
Flask-Cors==3.0.10
fonttools==4.29.1
frozenlist==1.4.0
greenlet==3.0.0
icalendar==4.0.7
idna==3.2
ipykernel==6.16.2
ipython==8.5.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.3.2
langchain==0.0.314
langsmith==0.0.43
lxml==4.8.0
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
multitasking==0.0.10
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.22.2
openai==0.28.1
packaging==21.3
pandas==1.4.1
pandas-datareader==0.10.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.3
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.29.0
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.2
stack-data==0.5.1
superpowered-sdk==0.1.22
tenacity==8.2.3
tomli==2.0.1
tornado==6.2
tqdm==4.66.1
traitlets==5.5.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
yarl==1.9.2
yfinance==0.1.70
